Added a 'hysteresis' filter for visibility and pointing detection
The (100% geometrical) filter works by adding margins to enter and quit
the visible/pointed zone.
It is 'harder' to start seeing/pointing an object, and also 'harder' to
stop seeing/pointing it.
This should prevent instable behaviour when visibility/pointing cones
are near objects.

See hri_visibility.cpp:1154 for details on the algorithm.

While here:
 - fixed identation in some parts
 - used correct enum (HRI_VISIBILITY_PLACEMENT) type for arguments
   instead of integers.